Create crontabfile
This script assumes that you have already set up an AWS Glue job to restore the database from the backup file. You'll need to replace the variables at the top of the script with your own values.

To run this script automatically on a schedule, you can add an entry to your crontab file. For example, to run the script every day at 3 AM, you could add the following line to your crontab.



This will run the script at 3 AM every day. 

Be sure to replace /path/to/restore_script.sh with the actual path to your script.